Restructured the modules
