+++ src/components/ChartTable/ChartTable.types.js
@@ -2,7 +2,8 @@ import moment from "moment";
 import numeral from "numeral";
 import type { Data, EnglandData } from "types/Data";
 
-declare type TableStructure = {
+
+declare export type TableStructure = {
 
     metadata: Array<{
         key: string,
@@ -23,7 +24,12 @@ declare type TableStructure = {
 } // TableStructure
 
 
-declare export type EnglandTableProps = { englandData: EnglandData, structure: TableStructure }
+declare export type EnglandTableProps = {
+
+    englandData: EnglandData,
+    structure: TableStructure
+
+} // EnglandTableProps
 
 
 export interface ChartTableProps {
@@ -38,3 +44,27 @@ export interface CharTableState {
     view: string
 
 } // CharTableState
+
+
+declare export type TitleOrDescriptionValues = {
+
+    totalCases: string,
+    dailyCases: string,
+    totalDeaths: string,
+    dailyDeaths: string
+
+} // TitleOrDescriptionValues
+
+
+declare export type TitleOrDescription = {
+
+    titles: TitleOrDescriptionValues,
+    descriptions: TitleOrDescriptionValues
+
+} // TitleOrDescription
+
+
+declare export type ChartsProps = TitleOrDescription & ChartTableProps
+
+
+declare export type TablesProps = TitleOrDescription & ChartTableProps
